+++ AddOns/TidyPlates_ThreatPlates/CHANGELOG.md
@@ -1,3 +1,5 @@
-# 11.2.13 (2024-06-29)
+# 12.0.0-beta4 (2024-07-21)
 
-* Added a delay when updating several CVars for soft targeting as they no longer can be changed in combat [Comment #7346].
+* Updated Threat Plates for The War Within pre-patch [GH-524, GH-525].
+* Added Bad Manner (Hunter, Pet) as a CC aura to the Auras widget [GH-531].
+* Added an option to change friendly and enemy healthbar sizes independently from each other (Retail Wow) [GH-459].